refactor: API response 타입 추가
