+++ src/apis/index.ts
@@ -1,7 +1,8 @@
 import { generateBasicToken } from "../utils/auth";
 import { API_URL, USER_ID, USER_PASSWORD } from "../constants/cart";
+import { CartItemType } from "../types/cart";
 
-export const getCartItems = async () => {
+export const getCartItems = async (): Promise<CartItemType[]> => {
   const token = generateBasicToken(USER_ID, USER_PASSWORD);
   const response = await fetch(`${API_URL}/cart-items`, {
     method: "GET",
